+++ projects/ngx-xchange-ui/src/components/list-item-selector/list-item-selector/list-item-selector.component.ts
@@ -10,6 +10,11 @@ export class ListItemSelectorComponent implements OnInit {
   @Input() label: string = '';
   @Input() checked: boolean = false;
   @Input() name: string = '';
+  @Input() completed: boolean = false;
+  @Input() userItem: boolean = false;
+  @Input() bigItem: boolean = false;
+  @Input() transparent: boolean = false
+  
     @Output() checkedChange = new EventEmitter<boolean>();
   
     toggleChequed() {