+++ projects/ngx-xchange-ui/src/directives/container/styles.ts
@@ -1,0 +1,27 @@
+export const APP_TAILWIND_STYLES: { [key: string]: string } = {
+    listSelector : `
+    relative
+    w-full 
+    flex 
+    flex-col 
+    gap-6
+    `,
+    assignmentCard:`
+    w-full
+    flex
+    flex-col
+    gap-4
+    bg-Grey050
+    p-4
+    items-start
+    rounded-lg
+    `,
+    rightSideModal:`
+    w-[55%] 
+    h-full 
+    bg-white 
+    rounded-r-2xl 
+    pt-12 
+    px-8
+    `
+}