+++ projects/ngx-xchange-ui/src/types.ts
@@ -1,0 +1,3 @@
+export interface ButtonType { 
+    [key:string]: string;
+  }